Store initial refresh token with every new access token
It looks like only the initial access token provides a refresh token.
All access tokens that are requested by using this refresh token, don't
provide an own refresh token. As we are always overwriting the old
access token with the new one, we must keep the old refresh token if the
new access token does not provide one by itself.