Use dedicated database instances for testing
Although the connection was configurable before, it couldn't be used
effectively for testing as the object behind the database was still the
same and thus could not simply be teared down and set up again for each
test.
Therefore, this change explicitly creates dedicated Database instances,
so there is no magic "db" object that is used in the background for
every transaction.
The drawback of this is, that the database must now be provided for each
transaction and such, every component working with the database needs to
know about it.