Add template filter to format timestamps as datetime objects
As we are currently unable to store datetime objects as-is in the json
field of Pony ORM, we must convert the stored timestamp objects to
datetime objects for better visualization. Formerly, this was done in
the view class before storing the dates in the template context.
When switching to JSON API views, we have a second layer in which we
would have to (de)serialize those objects and thus, we move that task to
the very end of the pipeline - which is the template itself.

This might not be the most performant solution, but it does the
convertion in the place where it's needed in the end.