Combine modules on index page via templates without iframes
This changes the way the module data is provided for the index page
template and avoids the usage of iframes.

This allows a simpler styling of the page as everything is part of a
single HTML document and thus, CSS classes and Bootstrap will work
togetherin a better way. The drawback of this solution is that the
templates (more precisely the template context) of each module is not
independent of the others any more (which was the case with iframes as
every document was a single HTML page rendered via a dedicated flask
endpoint).