Retrieve module data in generic function (not view specific)
Replaces the former requests.get() call in the index view with a
"normal" method call. This way we are not relying on any
connection/host/port being accessible from the outside (via requests).

The ApiViews are still in place and just use the get_module_data().

This should also make the whole application testable via Flask's
test_client (which does not work when relying on internal GET requests
with python's requests module.