Use gunicorn to make flirror test app reachable by Backstop JS process
Originally, I thought we can use flask's testing server, but this seems
to be only accessible from within python and not from the outside (e.g.
via a requests call or Backstop JS). This is most probably, because it
only serves the uwsgi application code, but not a uwsgi server.