Use flask development server (flask:run) rather than gunicorn
It stated out that using gunicorn (or any other uwsgi server) has the
drawback that we can't use freezegun to freeze the timestamps. Thus, the
relative timestamps shown in the flirror modules are depending on when
the test is executed (which is bad). To be able to use freezegun, we
must start the flirror application programatically from within python
and not via a subprocess. This is possible with flask's development
server which can be invoked programatically by calling the run_command()
function.

When calling flask's run command, both click applications
(run-backstop.py and flask) seem to interfere with each other resulting
in some very awkward "unexpected argument" errors. I tried different
version like calling the function directly or via click, but everything
resulted in the same error.

Now, I'm calling the underlying werkzeug function directly and mock all
necessary parameters (resp. look those up from flask):

The source of everything is
https://github.com/pallets/flask/blob/master/src/flask/cli.py#L829