Execute CSS regression tests inside Docker
When trying to execute the CSS regression tests in Travis CI, the
builds where always failing. It took me quite some time to find out that
the tests were failing because the font is rendered differently on mac
OS and Linux. So I thought the font I'm using for flirror is not
available on the build node and thus included it directly from the
Google APIs.

As this didn't help I did some research and found an issue [1] that
described exactly this problem. As it stated out, the anti-aliasing is
working differently on mac OS and Linux, leading to fonts that are
rendered slightly different.

After I found that, I found the relevant entry in the Backstop JS
documentation [2] saying that one should execute Backstop JS within
docker to avoid testing across different environments.

Backstop JS already provides a simple --docker option which
automatically executes the backstop commadn within a docker container.
Unfortunately this solution doesn't work for us, as the docker container
won't have access to the flask development instance which is running on
http://localhost:5000. We could handle this by switching to
host.docker.internal in the backstop configuration file, but that would
also only work on Mac OS, not on Linux.

Long story short, to avoid having differnt configuration files or even
different processes on how the tests are run, I decided to build a
custom Dockerfile that inherits from the backstopjs/backstopjs
Dockerfile, installs flirror and runs the run-backstop.py script (which
already does alle the necessary test setup).

So, here it is: The first reference file that works both, locally and
within Travis CI. 🎉🎉🎉

[1] https://github.com/garris/BackstopJS/issues/361
[2] https://github.com/garris/BackstopJS#using-docker-for-testing-across-different-environments