Remove tox css environment
Since we are running our CSS regression tests now inside a Docker
container, there is no need for this tox env any longer.

Spinning up the docker container (once build) is also much faster than
spinning up the tox env every time (most probably because of the nodejs
installation - although it was already done during the to the creation
of the environment).