Make max number of stored newsfeed entries configurable
So far, we were storing all that were returned by the newsfeed, which
doesn't make much sense as we are currently only displaying the first
one.

In future it would be cool to slide between the entries of one newsfeed,
so making this configurable should be bettern than reducing to a single
entry result.