Don't include module templates (except clock) directly in index.html
First, this doesn't work with the new plugin mechanism, as the template
name is configured in the module itself.
Second, even with the current solution, a missing template wouldn't harm
anything as the content of the module's <div> is anyways reloaded via
ajax.

Thus, for the future, we could just get rid of those direct template
includes and let the ajax.html do its thing.

The clock module is kind of a special thing here. Currently, I have no
idea how we could make this work with the new plugin mechanism.