Load chart.js in index template rather than the module template
When the templates are not directly included in the index page, but
reloaded asynchronously via ajax, the stocks modules fails to load the
chart.js source file, saying "Chart not found".

Including the chart.min.js file directly in the layout template instead
of reloading it asynchronously seems to help.

For the long term, we should find a way to let modules specify
additional JS/CSS files instead of putting those directly in the layout
template.