Don't load any data in index page
With that, the index page will only ensure that the modules are
positioned properly. The content of each tile will be loaded
asynchronously via ajax.

The main reason for this is that I want to get rid of the
raw/template parameters in the module APIs and this is the only place
where the raw parameter is actually used.

Although this seems to be a step back and makes flirror depending on
Javascript in some way, it simplifies the overall workflow as there is
only one way left how content is loaded in the web UI and we don't have
to maintain proper template and raw endpoints for all modules.

Also, ajax is nothing new by now and should work properly on all
browsers ;-)