Discover and register plugins starting with "flirror_"
This uses the mechanism described in [1] to look up any installed
package starting with "flirror_". If those packages define either a
FLIRROR_MODULE or FLIRROR_MODULES variable, and their values are
pointing to a valid FlirrorModule instance (or a list of FlirrorModule
instances), we will load them as additional modules into the
flirror application.

[1]: https://packaging.python.org/guides/creating-and-discovering-plugins/