Always prefix a module endpoint with its name
This avoids name conflicts between different modules and we can drop the
URL lookup in the ajax reloading template as we can rely on the
module's name there as well.

This only works if each module provides it's view on route "/", so we
should define this as a requirement for the plugin mechanism.