Provide a decorator view() that ensures a modules url rule is always "/"
Previously, we were using Flask's route() decorator, but were relying on
the route being set to "/" to make the loading of the views in the index
page work.

To avoid wrong inputs for the route() decorator, we now provide our own
that ensures the correct route being set.

This also simplifies the modules API, as a module now simply provides a
view() and a crawler().