Rename config key modules.type to modules.module
This makes it more clear that one should specify the module's name here
and not some mysterious type.

To avoid having the name modules on different hierarchy levels in the
data structure, I renamed the "modules" list used in the template
context to "tiles" (which is what they are in the end - from a visual
point of view).