Don't define a FLIRROR_OBJECT_KEY per module but use the module_id
The FLIRROR_OBJECT_KEY was used to identify the data stored for a
dedicated module in the database. However, always have to specify
this key makes the plugin mechanism more complex. We could get rid of
this and just use the module's id instead. This should be unique enough
for a single flirror instance.

If needed, one could now also specify an additional object_key when
storing data. The default will always be "data". This is helpful if a
module stores more than a single dataset in the database.