Prepuild the flirror-css image and upload it to Docker hub
With this change we can use the prebuilt flirror-css image from docker
hub for our Travis CI builds. To test the current development version
we mount the local workspace into the /src directory of the container.
This should speed up the Travis CI builds for the CSS tests as we don't
have to build the Docker image with every build.