Improve css-Dockerfile size
When pushing the flirror-css Dockerfile to Dockerhub I recognized its
filesize for the first time. Thus, I decided to build a custom
Docker image for the BackstopJS base rather than using the official
BackstopJS image. This custom image does mostly the same like the
official one, but it uses a debian-slim image as base and does a little
more cleanup.

With that change we can now also specify the BackstopJS and NodeJS
version in our custom Dockerfile.